Fixed a problem with the value changing of a pointer
I need to find a better solution for updating the value of the pointer, the start is to pass a double pointer to the function that requires the pointer, but I require at least 3 levels, and passing double pointer to all functions is not good for me. Maybe this is the best solution.